Search within a category and more logic term
This basically just allows the user to have a search input field within
a category to search that certain category for ads. It’s really fast
and useful to have handy.
Also, another enhancement is just that I changed Price to Price Range,
because it makes a little more sense